server/requirements.txt
django==2.2
djangorestframework
requests
djoser
requests_oauthlib

# linting
pylint
pylint-django